Place XHTML elements in the correct namespace.
XHTML elements are required to be in a namespace in order to be recognized
as part of XHTML.  Without the namespace declaration, browsers render pages
as a series of XML nodes, which is not very useful.  Add this declaration
for XHTML documents.  Also, adjust the xmldoc_from_string test function to
parse the document as an XML document if an XML namespace declaration is
present.