resolves #651 First version of the manpage backend
This is just a copy of the html backend with only enough edits to build
the asciidoctor manpage as closely as needed to the output of the a2x
command.

This commit doesn't clear any unused sections in the hopes that if you
need them, you do what I did and modify them in a per needed basis.