Fix image icons for long callouts in HTML5 backend
When the `icons` attribute is set to "image" with the html5 backend,
then the callout icons in callout table shrink and then disappear as any
of the callout text wraps to more than one line.

The fix is to set the 'max-width' CSS property of icon img's to 'none'
as is done for admonitionblock icons. Following the example of
admonitionblock, this patch adds an 'icon' class to the <td> element
containing the callout icon and a corresponding line in the default
stylesheet.

Fixes issue #1895