CSS: use bold for menus.
Menus seem to be a lot easier to read with `font-weight:bold`. Buttons are already bold, but in that case it is implemented by using a `<b>` tag in the HTML. I suppose CSS or the HTML `<strong>` tag would be more suitable in both cases.

This commit adds `.menuseq,.menu{font-weight:bold}` to asciidoctor-default.css.