manpage: use monospaced font for literal and listing blocks
Most of man page viewing happens on the terminal, which
is monospaced anyway, but one can generate a PDF or PostScript
of a man page as well, for example:

	man -Tps man > man.ps

In this case, if a man page was generated by asciidoctor, all the
blocks that are supposed to be using monospaced font, won't.

Fix is easy.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>