don't apply passthrough substitution inside stem macros
I just moved processing of the stem substitutions before processing of
the passthrough substitutions. It seems that it works well and the side
effects of this change are acceptable, i.e. they will probably affect
much less users than the current undesirable behaviour of + signs inside
the stem macros.

Fixes #3409