Add a labelspan block attribute, so man pages produced with html5 looks more like real manpages.
There are already two other block atributes named labelwidth and itemwidth that are integers expressed in %, the label width define the width of so called item area for description list (dlist) while labelwidth defines the width of what so called the content support. This patch define a new block attributes called labelspan, it is an integer that provide hints on number of char  that can fit in the label column (labelwidht), if the text length we will enter in the label cell won't fit in, we simply say we want to span this cell to then next one, and bring the item cell on the next line, i.e next raw in <table> parlance.