reader: warn when not seeing closing endif
It's quite easy to leave out an endif and then have the rest of the
document accidentally be skipped.  This implements a small check when
skipping lines in the pre-processor; if we are currently in skipping
mode, and we can see the end of the line stack, and it doesn't include
an endif, give the user a warning that they've forgotten to close
their conditional.

Closes: https://github.com/asciidoctor/asciidoctor/issues/3545